+++ src/bgen/Extensions/ExtensionMethods.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Xml;
+
 using Foundation;
 using ObjCRuntime;
 
@@ -18,6 +20,18 @@ public static class GeneratorExtensions {
 		return sw;
 	}
 
+	public static void LoadWithoutNetworkAccess (this XmlDocument doc, string filename)
+	{
+		using (var fs = new FileStream (filename, FileMode.Open, FileAccess.Read)) {
+			var settings = new XmlReaderSettings () {
+				XmlResolver = null,
+				DtdProcessing = DtdProcessing.Parse,
+			};
+			using (var reader = XmlReader.Create (fs, settings)) {
+				doc.Load (reader);
+			}
+		}
+	}
 }
 
 public static class ReflectionExtensions {
